fix: nav바 관심사 분리
